Refactored serde into a package. (#1100)
* Refactored serde into a package.

Removed `dump_json`, use pickle directly instead.

Implemented encoding for `tuple` and `set`, they were encoded as lists before.

* Renamed _code to _repr.

* Removed print-debug.

* Moved Stateful/Statelesl to serde.

* Black.

* Support for 3.6

* Added missing re-import.

* Py36 support.

* Removed example from docstring to pass test.

* Make doctest happy.
